requirements.txt
pyyaml
pandas
pdfplumber